chore: remove subdirectory mc-driver that is useless
