+++ go/internal/multipeer-connectivity-transport/driver/MCManager.h
@@ -1,6 +1,6 @@
 //
 //  MCManager.h
-//  mc-driver
+//  driver
 //
 //  Created by Rémi BARBERO on 31/03/2020.
 //  Copyright © 2020 Rémi BARBERO. All rights reserved.
