+++ go/internal/multipeer-connectivity-transport/driver/MCManager.m
