--- go/internal/multipeer-connectivity-transport/driver/mc-driver/cgo_bridge_unsupported.go
@@ -1,3 +1,0 @@
-// +build !darwin !cgo
-
-package mcdriver
